+++ repka/api.py
@@ -63,8 +63,8 @@ class BaseRepository(ConnectionMixin, Generic[T]):
     def table(self) -> Table:
         pass
 
-    def serialize(self, entity: T, **kwargs: Any) -> Dict:
-        return model_to_primitive(entity, without_id=True, **kwargs)
+    def serialize(self, entity: T) -> Dict:
+        return model_to_primitive(entity, without_id=True)
 
     @abstractmethod
     def deserialize(self, **kwargs: Any) -> T:
@@ -73,7 +73,7 @@ class BaseRepository(ConnectionMixin, Generic[T]):
     @property
     def ignore_insert(self) -> Sequence[str]:
         """
-        Columns that will be ignored on insert while serialization
+        Columns will be ignored on insert while serialization
         These columns will be set after insert
 
         See following tests for example:
@@ -83,15 +83,17 @@ class BaseRepository(ConnectionMixin, Generic[T]):
         return []
 
     async def insert(self, entity: T) -> T:
+        # key should be removed manually (not in .serialize) due to compatibility
+        serialized = {
+            key: value
+            for key, value in self.serialize(entity).items()
+            if key not in self.ignore_insert
+        }
         returning_columns = (
             self.table.c.id,
             *(getattr(self.table.c, col) for col in self.ignore_insert),
         )
-        query = (
-            self.table.insert()
-            .values(self.serialize(entity, exclude=self.ignore_insert))
-            .returning(*returning_columns)
-        )
+        query = self.table.insert().values(serialized).returning(*returning_columns)
 
         rows = await self.connection.execute(query)
         row = await rows.first()